Potential fix for code scanning alert no. 7: Multiplication result converted to larger type

+++ Application/src/grabber/gif.h
@@ -351,7 +351,7 @@ void GifMakePalette( const uint8_t* lastFrame, const uint8_t* nextFrame, uint32_
     
     // SplitPalette is destructive (it sorts the pixels by color) so
     // we must create a copy of the image for it to destroy
-    int imageSize = width*height*4*sizeof(uint8_t);
+    size_t imageSize = static_cast<size_t>(width) * static_cast<size_t>(height) * 4u * sizeof(uint8_t);
     uint8_t* destroyableImage = (uint8_t*)GIF_TEMP_MALLOC(imageSize);
     memcpy(destroyableImage, nextFrame, imageSize);
